requirements.txt
flask
beautifulsoup4
celery
redis
requests
pymongo
pyaml
